Serde refactoring
